requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.3.4
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
beautifulsoup4==4.10.0
certifi==2019.3.9
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==3.4.7
cuda-python==11.6.0
cycler==0.10.0
Cython==0.29.28
datasets==1.18.4.dev0
decorator==5.1.1
defusedxml==0.6.0
dill==0.3.4
Django==2.2.14
django-cors-headers==3.7.0
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-rest-knox==4.1.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
docopt==0.6.2
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.1.0
graphics.py==5.0.1.post1
huggingface-hub==0.4.0
idna==2.8
importlib-metadata==4.10.0
jiwer==2.3.0
joblib==1.1.0
Js2Py==0.71
kiwisolver==1.3.1
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
matplotlib==3.4.2
multidict==6.0.2
multiprocess==0.70.12.2
numba==0.55.1
numpy==1.20.3
oauthlib==3.0.1
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pipwin==0.5.1
pooch==1.6.0
pyarrow==7.0.0
PyAudio @ file:///C:/Users/sntsh/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycodestyle==2.5.0
pycparser==2.20
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.1
python-decouple==3.1
python-Levenshtein==0.12.2
python3-openid==3.1.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.1.18
requests==2.21.0
requests-oauthlib==1.2.0
resampy==0.2.2
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.6.3
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
sqlparse==0.2.4
stripe==2.27.0
threadpoolctl==3.1.0
tokenizers==0.11.5
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.62.3
transformers @ git+https://github.com/huggingface/transformers.git@2c2a31ffbcfe03339b1721348781aac4fc05bc5e
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.24.2
Wave==0.0.2
xxhash==2.0.2
yarl==1.7.2
zipp==3.7.0
